requirements.txt
alembic==1.7.6
aniso8601==9.0.1
api==0.0.7
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.1
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
Flask==2.0.3
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RBAC==0.5.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
nose==1.3.7
packaging==21.3
paramiko==2.10.1
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.0.1
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
texttable==1.6.4
tomli==2.0.1
urllib3==1.26.8
websocket-client==0.59.0
Werkzeug==2.0.3
